Extract TSDocParser._parseLines() into a separate LineExtractor file
